cpuset_memory_spread_test:Modify the creation of fifo files
When the cgroup file system cannot be mounted,
the test will be skipped directly in the setup() of cpuset_funcs.sh.
This will cause the created myfifo file to not be deleted.
When the use case is run again,
an error will be reported that myfifo already exists.

Therefore, myfifo can be created in setup() of cpuset_funcs.sh,
so that it can be naturally created in setup() and deleted in cleanup().